Refactored to fit in the template style and refactored to be cleaner / easier to read.
